+++ src/rofi_rbw/selector/wofi.py
@@ -45,7 +45,7 @@ class Wofi(Selector):
         self,
         entry: DetailedEntry,
         show_help_message: bool,
-        keybindings: dict[str, Action],
+        keybindings: list[Keybinding],
         additional_args: list[str],
     ) -> tuple[list[Target] | None, Action | None]:
         parameters = ["wofi", "--dmenu", "-p", "Choose target", *additional_args]